correct IRI for Mississipi_River
Technically the IRI <http://dbpedia.org/page/Mississippi_River> does not identify the river, it identifies *the web page describing* the river. The correct IRI for the river is <http://dbpedia.org/resource/Mississippi_River> (and yes, it redirects to the 'page' URL, because obviously, you can not retrieve the river itself over the internet...).